+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- GUI: Support outbound connections to makers through libp2p circuit relays.
+
 ## [4.14.0] - 2026-08-22
 
 - ASB: The `get-swaps` RPC response now includes the Bitcoin redeem address for each swap.